% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documantation.R
\docType{data}
\name{R88}
\alias{R88}
\title{Table R88}
\format{
A data frame with 2 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{MahMorajeh}:
\itemize{
\item Farsi name: \emph{ماه مراجعه به خانوار}
\item Description: The month of visiting the household
}
}
}
\usage{
R88
}
\description{
{R for Rural} \cr
{88 for 1388}
}
\keyword{datasets}
