% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documantation.R
\docType{data}
\name{R88P1}
\alias{R88P1}
\title{Table R88P1 \cr
Member characteristics of rural household in 1388 \cr
خصوصيات اجتماعي اعضاي خانوار روستایی سال 1388}
\format{
A data frame with 10 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{شماره ردیف اعضای خانوار}
\item Description: Row number of family members
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{بستگی با سرپرست}
\item Description: Relationship with the guardian
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{جنس}
\item Description: Gender
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{سن}
\item Description: Age
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{وضع سواد}
\item Description: Literacy status
}
\item \code{DYCOL07}:
\itemize{
\item Farsi name: \emph{آیا در حال حاضر تحصیل می کند؟}
\item Description: Is she/he currently studying?
}
\item \code{DYCOL08}:
\itemize{
\item Farsi name: \emph{پایه یا مدرک}
\item Description: Education degree
}
\item \code{DYCOL09}:
\itemize{
\item Farsi name: \emph{وضع فعالیت}
\item Description: Employment status
}
\item \code{DYCOL10}:
\itemize{
\item Farsi name: \emph{وضع زناشویی}
\item Description: Marital status
}
}
}
\usage{
R88P1
}
\description{
{R for Rural} \cr
{88 for 1388} \cr
{P for Part}
}
\keyword{datasets}
